refactor: MoviesContainer를 이용하여 중복 코드 제거
