+++ repeat/step2/src/components/base/index.ts
@@ -1,4 +1,5 @@
 export * from "./Header";
 export * from "./Movies";
+export * from "./MoviesContainer";
 export * from "./NotFoundMovies";
 export * from "./Message";
